client/requirements.txt
pillow>=10.0.0
pyperclip>=1.8.2
requests>=2.31.0
